[compile] fix warning.

+++ design_mode/cpp/16_iterator.cpp
@@ -19,7 +19,7 @@ public:
 class Aggregate {
 public:
   virtual void Push(const string& str) = 0;
-  virtual string Pop(int index) = 0;
+  virtual string Pop(size_t index) = 0;
   virtual size_t Count() = 0;
   virtual Iterator* CreateIterator() = 0;
 };
@@ -76,7 +76,7 @@ public:
     vec_str_.push_back(str);
   }
   
-  string Pop(int index) {
+  string Pop(size_t index) {
     return vec_str_[index];
   }
